added focus to text input and click to due date in maintenance

+++ app/src/main/java/pl/kamilbaziak/carcostnotebook/ui/cardetailsfeagment/petroltab/TankFillAdapter.kt
@@ -51,7 +51,7 @@ class TankFillAdapter(
             bindining.apply {
                 textOdometer.text = root.context.getString(
                     R.string.odometer_item_value,
-                    tankFill.quantity,
+                    tankFill.quantity.toDouble(),
                     unit.shortcut()
                 )
                 textDate.text = tankFill.created.toDate()
